+++ Package.swift
@@ -699,10 +699,10 @@ let package = Package(
     products: [
         .library(
             name: "RealmCore",
-            targets: ["Core"]),
+            targets: ["RealmCore"]),
         .library(
             name: "RealmQueryParser",
-            targets: ["QueryParser"]),
+            targets: ["RealmQueryParser"]),
         .library(
             name: "RealmCapi",
             targets: ["Capi"]),
@@ -718,7 +718,7 @@ let package = Package(
             publicHeadersPath: "."
         ),
         .target(
-            name: "Core",
+            name: "RealmCore",
             dependencies: ["Bid"],
             path: "src",
             exclude: ([
@@ -747,8 +747,8 @@ let package = Package(
                 .linkedLibrary("z"),
             ]),
         .target(
-            name: "QueryParser",
-            dependencies: ["Core"],
+            name: "RealmQueryParser",
+            dependencies: ["RealmCore"],
             path: "src/realm/parser",
             exclude: [
                 "CMakeLists.txt",
@@ -766,7 +766,7 @@ let package = Package(
             ] + cxxSettings),
         .target(
             name: "SyncServer",
-            dependencies: ["Core"],
+            dependencies: ["RealmCore"],
             path: "src",
             exclude: ([
                 "CMakeLists.txt",
@@ -789,7 +789,7 @@ let package = Package(
             cxxSettings: cxxSettings),
         .target(
             name: "Capi",
-            dependencies: ["Core", "QueryParser"],
+            dependencies: ["RealmCore", "RealmQueryParser"],
             path: "src/realm/object-store/c_api",
             exclude: [
                 "CMakeLists.txt",
@@ -808,7 +808,7 @@ let package = Package(
             path: "src/swift"),
         .target(
             name: "ObjectStoreTestUtils",
-            dependencies: ["Core", "SyncServer"],
+            dependencies: ["RealmCore", "SyncServer"],
             path: "test/object-store/util",
             exclude: [
                 "baas_admin_api.hpp",
@@ -824,7 +824,7 @@ let package = Package(
             ] + cxxSettings) as [CXXSetting]),
         .target(
             name: "ObjectStoreTests",
-            dependencies: ["Core", "QueryParser", "ObjectStoreTestUtils"],
+            dependencies: ["RealmCore", "RealmQueryParser", "ObjectStoreTestUtils"],
             path: "test/object-store",
             exclude: [
                 "CMakeLists.txt",